attempt to lazy-load split all controlls
