+++ src/controls/controls.css
@@ -7,10 +7,6 @@
   user-select: none;
 }
 
-.showAllToggle {
-  margin-bottom: 0;
-}
-
 .inlineControls {
   display: flex;
   flex-direction: row;
