+++ src/controls/show-charts-toggle.css
@@ -1,0 +1,3 @@
+.showAllToggle {
+  margin-bottom: 0;
+}
